Move the version 7 lock algorithm out of the free lockFlake() function
The `computeLocks` machinery is moved from the free function
`lockFlake()` in flake.cc into a new static member function
`LockedFlakeV7::lockFlake()` in lockfile-v7.cc. The free function now
only contains version-independent work, expressed in terms of the
abstract `LockedFlake` interface:

* Reading the old lock file into a `nlohmann::json`, and constructing
  a `LockedFlake` from it (via a new `LockedFlakeV7` constructor that
  takes the JSON). This is where we will dispatch on the lock file
  version in the future.

* Change detection, by comparing the virtual `toJSON()` serializations
  of the old and new locked flakes. This also does the right thing
  across future version migrations (a version change is always a
  change).

* Printing the diff, via a new abstract `LockedFlake::diff()` method
  that shows the differences relative to an older `LockedFlake`. The
  version 7 implementation diffs against an empty lock file if the old
  lock file is not version 7, so all inputs show up as added.

* Writing/committing the new lock file, using `to_string()` and the
  virtual `isUnlocked()`.

`LockedFlakeV7::lockFlake()` takes the old lock file as a `const
LockedFlake &` and downcasts it internally. The `LockFileV7` parsing
constructor now takes a `nlohmann::json` instead of the file contents,
and `getFlake()` with a lock root attribute path is exported from
flake.cc since the lock algorithm needs it from its new home.
